Add files via upload
How to Automatically Delete WooCommerce Products in WordPress

If you're running an online store using WooCommerce in WordPress, you might want to automatically delete products that are no longer relevant or in stock. This can help keep your website organized and improve performance by reducing the number of products that need to be loaded.

In this article, we'll show you how to create a custom WordPress plugin that automatically deletes WooCommerce products after a configurable interval.
How the Plugin Works

The plugin we'll create uses WordPress's built-in "get_posts" function to retrieve a list of WooCommerce products that match a certain criteria, such as a specified time since the product was last updated. The plugin then loops through the list of products and deletes each one using the "wp_delete_post" function.

The interval at which the plugin deletes products is configurable via a settings page in the WordPress admin area. This allows you to set the delete interval to a value that suits your needs.